% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processors.R
\name{post_processors_template_processing}
\alias{post_processors_template_processing}
\title{Template Processing}
\description{
Provides a way to specify templates in order to add the special tokens to each
input sequence as relevant.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{post_processors_template_processing$new()}}
\item \href{#method-clone}{\code{post_processors_template_processing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{post_processors_template_processing$new(
  single = NULL,
  pair = NULL,
  special_tokens = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{single}}{(Template) – The template used for single sequences}

\item{\code{pair}}{(Template) – The template used when both sequences are specified}

\item{\code{special_tokens}}{(Tokens) – The list of special tokens used in each sequences}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{post_processors_template_processing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
