% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokenizer.R
\name{tokenizer}
\alias{tokenizer}
\title{The tokenizer class}
\description{
The tokenizer class

The tokenizer class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tokenizer$new()}}
\item \href{#method-train}{\code{tokenizer$train()}}
\item \href{#method-encode}{\code{tokenizer$encode()}}
\item \href{#method-decode}{\code{tokenizer$decode()}}
\item \href{#method-get_vocab}{\code{tokenizer$get_vocab()}}
\item \href{#method-save}{\code{tokenizer$save()}}
\item \href{#method-token_to_id}{\code{tokenizer$token_to_id()}}
\item \href{#method-encode_batch}{\code{tokenizer$encode_batch()}}
\item \href{#method-enable_padding}{\code{tokenizer$enable_padding()}}
\item \href{#method-clone}{\code{tokenizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$new(model)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$train(files, trainer = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encode"></a>}}
\if{latex}{\out{\hypertarget{method-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$encode(
  sequence,
  pair = NULL,
  is_pretokenized = FALSE,
  add_special_tokens = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$decode(ids, skip_special_tokens = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vocab"></a>}}
\if{latex}{\out{\hypertarget{method-get_vocab}{}}}
\subsection{Method \code{get_vocab()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$get_vocab(with_added_tokens = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$save(path, pretty = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-token_to_id"></a>}}
\if{latex}{\out{\hypertarget{method-token_to_id}{}}}
\subsection{Method \code{token_to_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$token_to_id(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encode_batch"></a>}}
\if{latex}{\out{\hypertarget{method-encode_batch}{}}}
\subsection{Method \code{encode_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$encode_batch(
  inputs,
  is_pre_tokenized = FALSE,
  add_special_tokens = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enable_padding"></a>}}
\if{latex}{\out{\hypertarget{method-enable_padding}{}}}
\subsection{Method \code{enable_padding()}}{
Enable the padding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$enable_padding(
  direction = "right",
  pad_id = 0,
  pad_type_id = 0,
  pad_token = "[PAD]",
  length = NULL,
  pad_to_multiple_of = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{direction}}{(str, optional, defaults to right) – The direction in which
to pad. Can be either right or left}

\item{\code{pad_id}}{(int, defaults to 0) – The id to be used when padding}

\item{\code{pad_type_id}}{(int, defaults to 0) – The type id to be used when padding}

\item{\code{pad_token}}{(str, defaults to \verb{[PAD]}) – The pad token to be used when padding}

\item{\code{length}}{(int, optional) – If specified, the length at which to pad.
If not specified we pad using the size of the longest sequence in a batch.}

\item{\code{pad_to_multiple_of}}{(int, optional) – If specified, the padding
length should always snap to the next multiple of the given value.
For example if we were going to pad with a length of 250 but
\code{pad_to_multiple_of=8} then we will pad to 256.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
