% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trainers.R
\name{trainers_bpe}
\alias{trainers_bpe}
\title{Byte pair encoding trainer class}
\description{
Trainer capable of training a BPE model
}
\section{Super class}{
\code{\link[hftokenizers:Trainer]{hftokenizers::Trainer}} -> \code{trainer_bpe}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trainers_bpe$new()}}
\item \href{#method-clone}{\code{trainers_bpe$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trainers_bpe$new(
  vocab_size = NULL,
  min_frequency = NULL,
  show_progress = NULL,
  special_tokens = NULL,
  limit_alphabet = NULL,
  initial_alphabet = NULL,
  continuing_subword_prefix = NULL,
  end_of_word_suffix = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vocab_size}}{(int, optional) – The size of the final vocabulary, including all tokens and alphabet.}

\item{\code{min_frequency}}{(int, optional) – The minimum frequency a pair should have in order to be merged.}

\item{\code{show_progress}}{(bool, optional) – Whether to show progress bars while training.}

\item{\code{special_tokens}}{(List[Union\link{str, AddedToken}], optional) – A list of special tokens the model should know of.}

\item{\code{limit_alphabet}}{(int, optional) – The maximum different characters to keep in the alphabet.}

\item{\code{initial_alphabet}}{(List\link{str}, optional) – A list of characters to include in the initial alphabet, even if not seen in the training dataset. If the strings contain more than one character, only the first one is kept.}

\item{\code{continuing_subword_prefix}}{(str, optional) – A prefix to be used for every subword that is not a beginning-of-word.}

\item{\code{end_of_word_suffix}}{(str, optional) – A suffix to be used for every subword that is a end-of-word.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trainers_bpe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
