% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tokenizer.R
\name{tokenizer}
\alias{tokenizer}
\title{The tokenizer class}
\description{
The tokenizer class

The tokenizer class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tokenizer$new()}}
\item \href{#method-train}{\code{tokenizer$train()}}
\item \href{#method-encode}{\code{tokenizer$encode()}}
\item \href{#method-decode}{\code{tokenizer$decode()}}
\item \href{#method-get_vocab}{\code{tokenizer$get_vocab()}}
\item \href{#method-clone}{\code{tokenizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$new(model)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$train(files)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encode"></a>}}
\if{latex}{\out{\hypertarget{method-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$encode(
  sequence,
  pair = NULL,
  is_pretokenized = FALSE,
  add_special_tokens = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$decode(ids, skip_special_tokens = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vocab"></a>}}
\if{latex}{\out{\hypertarget{method-get_vocab}{}}}
\subsection{Method \code{get_vocab()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$get_vocab(with_added_tokens = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tokenizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
